+++ config/presentable.php
@@ -8,4 +8,12 @@ return [
      */
     'key' => 'presentable',
 
+    /**
+     * You can specify a set of default presentable classes that are always
+     * set on any class that uses the `IsPresentable` trait.
+     */
+    'defaults' => [
+//        'created_at' => App\Http\Presenters\CreatedAtPresenter::class,
+    ]
+
 ];